separate entities into modules
